start building release2 in CI

+++ release/build_release2.sh
@@ -8,36 +8,43 @@ export GIT_AUTHOR_EMAIL="user@comma.ai"
 
 export GIT_SSH_COMMAND="ssh -i /data/gitkey"
 
-# Create folders
-rm -rf /data/openpilot
-mkdir -p /data/openpilot
-cd /data/openpilot
+# set CLEAN to build outside of CI
+if [ ! -z "$CLEAN" ]; then
+  # Create folders
+  rm -rf /data/openpilot
+  mkdir -p /data/openpilot
+  cd /data/openpilot
 
-# Create git repo
-git init
-git remote add origin git@github.com:commaai/openpilot.git
-git fetch origin devel
+  # Create git repo
+  git init
+  git remote add origin git@github.com:commaai/openpilot.git
+  git fetch origin devel-staging
+else
+  git clean -xdf
+fi
+
 git fetch origin release2-staging
 git fetch origin dashcam-staging
 
-# Checkout devel
-#git checkout origin/devel
-#git clean -xdf
-
 # Create release2 with no history
-git checkout --orphan release2-staging origin/devel
+if [ ! -z "$CLEAN" ]; then
+  git checkout --orphan release2-staging origin/devel-staging
+else
+  git checkout --orphan release2-staging
+fi
 
 VERSION=$(cat selfdrive/common/version.h | awk -F\" '{print $2}')
 git commit -m "openpilot v$VERSION"
 
 # Build signed panda firmware
 pushd panda/board/
-cp -r /tmp/pandaextra /data/openpilot/
-RELEASE=1 make obj/panda.bin
+#cp -r /tmp/pandaextra /data/openpilot/
+#RELEASE=1 make obj/panda.bin
+make obj/panda.bin
 mv obj/panda.bin /tmp/panda.bin
 make clean
 mv /tmp/panda.bin obj/panda.bin.signed
-rm -rf /data/openpilot/pandaextra
+#rm -rf /data/openpilot/pandaextra
 popd
 
 # Build stuff
@@ -70,11 +77,15 @@ git commit --amend -m "openpilot v$VERSION"
 # Print committed files that are normally gitignored
 #git status --ignored
 
-# Push to release2-staging
-git push -f origin release2-staging
+if [ ! -z "$CI_PUSH" ]; then
+  # Push to release2-staging
+  #git push -f origin release2-staging
+  git push -f origin r2_staging_test
 
-# Create dashcam release
-git rm selfdrive/car/*/carcontroller.py
+  # Create dashcam release
+  git rm selfdrive/car/*/carcontroller.py
 
-git commit -m "create dashcam release from release2"
-git push -f origin release2-staging:dashcam-staging
+  git commit -m "create dashcam release from release2"
+  #git push -f origin release2-staging:dashcam-staging
+  git push -f origin release2-staging:d_staging_test
+fi
